Move tests for with into parser.
