+++ compiler/parser/src/lib.rs
@@ -33,4 +33,3 @@ mod python;
 mod context;
 mod string;
 pub mod token;
-mod with;
