--- compiler/parser/src/with.rs
@@ -1,61 +1,0 @@
-#[cfg(test)]
-mod tests {
-    use crate::parser::parse_program;
-
-    #[test]
-    fn test_with_statement() {
-        let source = "\
-with 0: pass
-with 0 as x: pass
-with 0, 1: pass
-with 0 as x, 1 as y: pass
-with 0 if 1 else 2: pass
-with 0 if 1 else 2 as x: pass
-with (): pass
-with () as x: pass
-with (0): pass
-with (0) as x: pass
-with (0,): pass
-with (0,) as x: pass
-with (0, 1): pass
-with (0, 1) as x: pass
-with (*a,): pass
-with (*a,) as x: pass
-with (0, *a): pass
-with (0, *a) as x: pass
-with (a := 0): pass
-with (a := 0) as x: pass
-with (a := 0, b := 1): pass
-with (a := 0, b := 1) as x: pass
-with (0 as a): pass
-with (0 as a,): pass
-with (0 as a, 1 as b): pass
-with (0 as a, 1 as b,): pass
-";
-        insta::assert_debug_snapshot!(parse_program(source, "<test>").unwrap());
-    }
-
-    #[test]
-    fn test_with_statement_invalid() {
-        for source in [
-            "with 0,: pass",
-            "with 0 as x,: pass",
-            "with 0 as *x: pass",
-            "with *a: pass",
-            "with *a as x: pass",
-            "with (*a): pass",
-            "with (*a) as x: pass",
-            "with *a, 0 as x: pass",
-            "with (*a, 0 as x): pass",
-            "with 0 as x, *a: pass",
-            "with (0 as x, *a): pass",
-            "with (0 as x) as y: pass",
-            "with (0 as x), 1: pass",
-            "with ((0 as x)): pass",
-            "with a := 0 as x: pass",
-            "with (a := 0 as x): pass",
-        ] {
-            assert!(parse_program(source, "<test>").is_err());
-        }
-    }
-}
